改为ARC；网络请求及文件下载换成精简版AFN；
改为ARC；网络请求及文件下载换成精简版AFN；

+++ ExceedMVC/Core/FileManager/FileManager.h
@@ -14,7 +14,7 @@
 + (NSString *)cachePathForFile;
 
 // 查看缓存占用空间（MB）
-+ (CGFloat)cacheUsed;
++ (float)cacheUsed;
 
 // 清理缓存
 + (void)clearAllCache;
